add MainView component
